# [helper] Avoid expensive calls in Builder constructors.