--- a/helper/tests/helper-tests-18/src/jmh/java/com/linkedin/avroutil1/compatibility/avro18/NewRecord.java
+++ b/helper/tests/helper-tests-18/src/jmh/java/com/linkedin/avroutil1/compatibility/avro18/NewRecord.java
@@ -158,4 +158,9 @@ public class NewRecord {
   public SpecificRecord processed110Builder() {
     return under110wbuilders.SimpleRecord.newBuilder().setStringField(MESSAGE).build();
   }
+
+  @Benchmark
+  public SpecificRecord processed111Builder() {
+    return under111wbuilders.SimpleRecord.newBuilder().setStringField(MESSAGE).build();
+  }
 }