The introduction of a `private Builder(boolean)` constructor is ugly,
but it's the simplest way I found to deal with this issue. It's private
and not exposed, so the ugliness shouldn't bleed into anything else.

Fixes #220. In addition, since 1.11 codegen can now be processed into
being compatible with older runtimes, add some benchmark methods that
weren't possible before.

Here's the benchmark output from my laptop, only the interesting bits
(the parts that changed):

Before:
```
$ ./gradlew :helper:tests:helper-tests-110:jmh
...
NewRecord.processed110Builder  thrpt   10    1307.549 ±    282.097  ops/ms
NewRecord.processed17Builder   thrpt   10    1070.114 ±    259.597  ops/ms
NewRecord.processed18Builder   thrpt   10    1231.135 ±    204.276  ops/ms
NewRecord.processed19Builder   thrpt   10    1338.624 ±    267.086  ops/ms
```

After:
```
$ ./gradlew :helper:tests:helper-tests-110:jmh
...
Benchmark                       Mode  Cnt       Score        Error   Units
NewRecord.processed110Builder  thrpt   10   29914.824 ±  20403.631  ops/ms
NewRecord.processed17Builder   thrpt   10   33792.257 ±  18056.973  ops/ms
NewRecord.processed18Builder   thrpt   10   33543.303 ±  19317.861  ops/ms
NewRecord.processed19Builder   thrpt   10   30624.449 ±  14613.924  ops/ms
```

Roughly a 25x improvement.